// src/enums/Commands.ts
/** Command byte codes for the charger protocol. */
export enum CommandEnum {
  BD380_QUERY_STATUS = 0x71,
  BD380_QUERY_SYSTEM = 0x73,
  BD380_SET_SYSTEM = 0x72,
  BD380_START_WORK = 0x70,
  BD380_VERIFY_PASSWORD = 0x74,
  EDIT_PASSWORD = 0x56,
  GET_CHARGE_NUMBER = 0x50,
  INFO = 0x57,
  NC_DATA = 0x91,
  QR_QUERY_CHARGE = 0x08,
  QR_SEND_SN = 0x0a,
  QR_UPDATE_CHARGE = 0x09,
  QUERY_BASIC_INFO = 0x5f,
  QUERY_CHANNEL_STATUS = 0x55,
  QUERY_NC2200_SETTING_PARAMETERS = 0x93,
  QUERY_SYSTEM_SETTING = 0x96,
  QUERY_SYSTEM_INFO = 0x5a,
  QUERY_VOLTAGE_INFO = 0x58,
  SEND_SN = 0x04,
  SET_SNC2200_SETTING_PARAMETERS = 0x94,
  SET_SNC2200_SYSTEM = 0x95,
  SET_SYSTEM_INFO = 0x11,
  SET_TURBO = 0x92,
  SNC2200_RESET = 0x97,
  START_CHARGE = 0x05,
  STOP_CHARGE = 0xfe,
  UPGRADE = 0x61,
  UPGRADE_OVER = 0x62,
  UPGRADE_PREPARE = 0x60,
  CHECK_NEW_VERSION = 0x65,
  CHECK_BOOT_VERSION = 0x66,
  QUERY_DC_STATUS = 0x75,
  SET_DC_STATUS = 0x76,
  SET_TEMPERATURE = 0x99,
}