// src/enums/SettingTypes.ts
export enum SettingType {
  RECYCLE = 0x00,
  TIME = 0x01,
  CAPACITY = 0x02,
  SOUND = 0x03,
  VOLTAGE_LOW = 0x04,
  TEMPERATURE = 0x05,
  BALANCED = 0x06,
  POWER = 0x07,
  DC_ENABLE = 0x08,
  DC_POWER = 0x09,
  DC_VOLTAGE = 0x0a,
  DC_CURRENT = 0x0b,
  LCD = 0x10,
  WARNING = 0x11,
  SLEEP_TIME = 0x12,
  RESET = 0x15,
}